requirements.txt
numpy
pandas
pyarrow
python-chess
requests
tqdm
torch